requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosqlite==0.20.0
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
keyring==25.7.0
MarkupSafe==3.0.3
more-itertools==10.8.0
multidict==6.7.0
propcache==0.4.1
pycparser==2.23
pylast==7.0.0
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
PySocks==1.7.1
python-vlc==3.0.21203
qasync==0.28.0
qt-material==2.17
requests==2.32.5
SecretStorage==3.5.0
shiboken6==6.10.1
sniffio==1.3.1
urllib3==2.5.0
yandex-music==2.2.0
yarl==1.22.0
yt-dlp==2025.11.12
ytmusicapi==1.11.2
